hayaoshiButton/requirements.txt
Flask
Flask-SocketIO
python-socketio[client] # WebSocketクライアント側も必要なら
python-engineio
bleak
Flask
Flask-SQLAlchemy
PyMySQL
